requirements.txt

viam-sdk

typing-extensions
